+++ tests/integrationtests/indicators/test_attribute_completeness.py
@@ -6,22 +6,27 @@ import geojson
 import plotly.graph_objects as pgo
 import plotly.io as pio
 import pytest
+from approvaltests import verify
 
 from ohsome_quality_api.indicators.attribute_completeness.indicator import (
     AttributeCompleteness,
 )
-from tests.integrationtests.utils import get_topic_fixture, oqapi_vcr
+from tests.integrationtests.utils import (
+    PytestNamer,
+    get_topic_fixture,
+    oqapi_vcr,
+)
 
 
 class TestPreprocess:
     @oqapi_vcr.use_cassette
-    def test_preprocess(
+    def test_preprocess_attribute_keys_single(
         self, topic_building_count, feature_germany_heidelberg, attribute_key
     ):
         indicator = AttributeCompleteness(
             topic_building_count,
             feature_germany_heidelberg,
-            attribute_key,
+            attribute_keys=attribute_key,
         )
         asyncio.run(indicator.preprocess())
         assert indicator.result.value is not None
@@ -29,13 +34,32 @@ class TestPreprocess:
         assert isinstance(indicator.result.timestamp_osm, datetime)
 
     @oqapi_vcr.use_cassette
-    def test_preprocess_multiple_attribute_keys(
+    def test_preprocess_attribute_keys_multiple(
         self, topic_building_count, feature_germany_heidelberg, attribute_key_multiple
     ):
         indicator = AttributeCompleteness(
             topic_building_count,
             feature_germany_heidelberg,
-            attribute_key_multiple,
+            attribute_keys=attribute_key_multiple,
+        )
+        asyncio.run(indicator.preprocess())
+        assert indicator.result.value is not None
+        assert isinstance(indicator.result.timestamp, datetime)
+        assert isinstance(indicator.result.timestamp_osm, datetime)
+
+    @oqapi_vcr.use_cassette
+    def test_preprocess_attribute_filter(
+        self,
+        topic_building_count,
+        feature_germany_heidelberg,
+        attribute_filter,
+        attribute_title,
+    ):
+        indicator = AttributeCompleteness(
+            topic_building_count,
+            feature_germany_heidelberg,
+            attribute_filter=attribute_filter,
+            attribute_title=attribute_title,
         )
         asyncio.run(indicator.preprocess())
         assert indicator.result.value is not None
@@ -44,19 +68,31 @@ class TestPreprocess:
 
 
 class TestCalculation:
-    @pytest.fixture(scope="class")
+    @pytest.fixture(
+        scope="class",
+        params=(
+            ["height"],
+            {
+                "attribute_filter": "height=* or building:levels=*",
+                "attribute_title": "Height",
+            },
+        ),
+    )
     @oqapi_vcr.use_cassette
-    def indicator(
-        self,
-        topic_building_count,
-        feature_germany_heidelberg,
-        attribute_key,
-    ):
-        i = AttributeCompleteness(
-            topic_building_count,
-            feature_germany_heidelberg,
-            attribute_key,
-        )
+    def indicator(self, request, topic_building_count, feature_germany_heidelberg):
+        if isinstance(request.param, list):
+            i = AttributeCompleteness(
+                topic_building_count,
+                feature_germany_heidelberg,
+                attribute_keys=request.param,
+            )
+        else:
+            i = AttributeCompleteness(
+                topic_building_count,
+                feature_germany_heidelberg,
+                attribute_filter=request.param["attribute_filter"],
+                attribute_title=request.param["attribute_title"],
+            )
         asyncio.run(i.preprocess())
         i.calculate()
         return i
@@ -68,9 +104,10 @@ class TestCalculation:
         assert indicator.result.description is not None
         assert isinstance(indicator.result.timestamp, datetime)
         assert isinstance(indicator.result.timestamp_osm, datetime)
+        verify(indicator.description, namer=PytestNamer())
 
-    @oqapi_vcr.use_cassette()
-    def test_no_features(self, attribute):
+    @oqapi_vcr.use_cassette
+    def test_no_features(self):
         """Test area with no features"""
         infile = os.path.join(
             os.path.dirname(os.path.abspath(__file__)),
@@ -95,26 +132,36 @@ class TestCalculation:
 
 
 class TestFigure:
-    @pytest.fixture(scope="class")
+    @pytest.fixture(
+        scope="class",
+        params=(
+            ["height"],
+            {
+                "attribute_filter": "height=* or building:levels=*",
+                "attribute_title": "Height",
+            },
+        ),
+    )
     @oqapi_vcr.use_cassette
     def indicator(
         self,
+        request,
         topic_building_count,
         feature_germany_heidelberg,
-        attribute_key,
     ):
-        indicator = AttributeCompleteness(
-            topic_building_count,
-            feature_germany_heidelberg,
-            attribute_key,
-        )
+        if isinstance(request.param, list):
+            indicator = AttributeCompleteness(
+                topic_building_count, feature_germany_heidelberg, request.param
+            )
+        else:
+            indicator = AttributeCompleteness(
+                topic_building_count,
+                feature_germany_heidelberg,
+                attribute_filter=request.param["attribute_filter"],
+                attribute_title=request.param["attribute_title"],
+            )
         asyncio.run(indicator.preprocess())
         indicator.calculate()
-        assert indicator.description == (
-            '28.8% of all "building count" features (all: 29936 elements)'
-            " in your area of interest have the selected additional attribute"
-            " height of buildings (matched: 8630 elements). "
-        )
         return indicator
 
     # comment out for manual test
@@ -129,7 +176,7 @@ class TestFigure:
         pgo.Figure(indicator.result.figure)  # test for valid Plotly figure
 
 
-def test_create_description():
+def test_create_description_attribute_keys_single():
     indicator = AttributeCompleteness(
         get_topic_fixture("building-count"),
         "foo",
@@ -139,14 +186,10 @@ def test_create_description():
     indicator.absolute_value_1 = 10
     indicator.absolute_value_2 = 2
     indicator.create_description()
-    assert indicator.description == (
-        '20.0% of all "building count" features (all: 10 elements) in your area of '
-        "interest have the selected additional attribute height of buildings "
-        "(matched: 2 elements). "
-    )
+    verify(indicator.description, namer=PytestNamer())
 
 
-def test_create_description_multiple_attributes():
+def test_create_description_attribute_keys_multiple():
     indicator = AttributeCompleteness(
         get_topic_fixture("building-count"),
         "foo",
@@ -156,11 +199,21 @@ def test_create_description_multiple_attributes():
     indicator.absolute_value_1 = 10
     indicator.absolute_value_2 = 2
     indicator.create_description()
-    assert indicator.description == (
-        '20.0% of all "building count" features (all: 10 elements) in your area of '
-        "interest have the selected additional attributes height of buildings, house "
-        "number, street address (matched: 2 elements). "
+    verify(indicator.description, namer=PytestNamer())
+
+
+def test_create_description_attribute_filter(attribute_filter, attribute_title):
+    indicator = AttributeCompleteness(
+        get_topic_fixture("building-count"),
+        "foo",
+        attribute_filter=attribute_filter,
+        attribute_title=attribute_title,
     )
+    indicator.result.value = 0.2
+    indicator.absolute_value_1 = 10
+    indicator.absolute_value_2 = 2
+    indicator.create_description()
+    verify(indicator.description, namer=PytestNamer())
 
 
 @pytest.mark.parametrize(