+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
-## Current Main 
+## Current Main
+
+### New Features
+
+- Support custom attribute definition via ohsome filter query for the Attribute Completeness indicator ([#848])
+
+[#848]: https://github.com/GIScience/ohsome-quality-api/pull/848
 
 
 ## Release  1.7.0